// references/math-raiders/client/src/data/mathFacts.ts
type OperationVariant = Infer<typeof Operation>;

export interface MathFact {
  left: number;
  right: number;
  operation: OperationVariant;
  grades: number[];
  tracks: string[];
}